requirements.txt
yt-dlp
pytube
pillow
spotipy
aiohttp
aiofiles
tgcrypto
py-tgcalls
python-dotenv
youtube-search-python
PyrogramMod @ https://github.com/PyrogramMod/PyrogramMod/archive/refs/heads/main.zip
